requirements.txt
flask
python-dotenv
openai
apscheduler
transformers